Teleporter Addition + others
I cleaned up the main menu screen, added a few things to the main game. teleporter is still being worked on, incomplete due to the fact that we still need to implement a timer, increasing difficulty based off timer, a difficulty selector and small QoL that are similar to risk of rain 2.