A* Tracking + Character controller
due to the map being way too big for navMesh, i decided to go with A* tracking for the enemies. I'm still working on the player controller, but what needs to be fixed is the jumping and enemy movement, along with the flow of levels (how it goes to the next level).